refactor(assembly): split topology families into modules
Extract Maven, Gradle, uv, Go, Pixi, and Hackage topology logic from the
monolithic topology.rs into family modules matching existing workspace merges.
