+++ src/assembly/mod.rs
@@ -17,15 +17,19 @@ mod conda_rootfs_merge;
 mod dart_workspace_merge;
 mod debian_source_merge;
 pub mod file_ref_resolve;
+mod go_workspace;
+mod gradle_multiproject;
 mod hackage_merge;
 mod huggingface_merge;
 mod ivy_dependencies_properties_assign;
+mod maven_reactor;
 mod mix_umbrella_merge;
 mod nested_merge;
 mod nix_flake_compat_merge;
 mod npm_resource_assign;
 mod npm_workspace_merge;
 mod nuget_cpm_resolve;
+mod pixi_topology;
 mod project_dependency_assign;
 mod python_requirements_assign;
 mod resource_assign;
@@ -33,6 +37,7 @@ mod ruby_resource_assign;
 mod sibling_merge;
 mod swift_merge;
 mod topology;
+mod uv_workspace;
 mod windows_update_merge;
 
 use std::collections::{BTreeSet, HashMap, HashSet};
